requirements.txt
absl-py==0.7.0
astor==0.7.1
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.7.1
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.18.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
numpy==1.16.1
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
pylint==1.9.4
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
